% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{fit_age_model}
\alias{fit_age_model}
\title{Fit age-dependent copper model}
\usage{
fit_age_model(data, cu_var = "Cu", age_var = "age.months")
}
\arguments{
\item{data}{Data frame containing copper and age}

\item{cu_var}{Name of copper variable (µg/L)}

\item{age_var}{Name of age variable (months)}
}
\value{
An object of class 'cu_age_model'
}
\description{
Fits a log-linear model of serum copper as a function of age.
}
