% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{fit_inflammation_score}
\alias{fit_inflammation_score}
\title{Fit inflammation score using PLS}
\usage{
fit_inflammation_score(
  age_model,
  data,
  fibrinogen_col = "fibrinogen",
  crp_col = "CRP",
  esr_col = "ESR",
  cu_var = "Cu",
  age_var = "age.months"
)
}
\arguments{
\item{age_model}{Object returned by fit_age_model()}

\item{data}{Data frame with inflammation markers and copper residuals}

\item{fibrinogen_col}{Name of fibrinogen column}

\item{crp_col}{Name of CRP column}

\item{esr_col}{Name of ESR column}

\item{cu_var}{Name copper variable}

\item{age_var}{Name of age variable (months)}
}
\value{
Object of class 'cu_inflammation_score'
}
\description{
Fit inflammation score using PLS
}
