% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{fit_reference_intervals}
\alias{fit_reference_intervals}
\title{Fit age-adjusted reference intervals for serum copper}
\usage{
fit_reference_intervals(age_model, data, n_boot = 1000)
}
\arguments{
\item{age_model}{Object returned by fit_age_model()}

\item{data}{Data used to estimate reference intervals}

\item{n_boot}{Number of bootstrap iterations}
}
\value{
An object of class 'cu_reference_intervals'
}
\description{
Estimates lower and upper reference limits using bootstrap residuals.
}
